+++ ampere/mirror.py
